Rename: 토론방 관련 클래스 및 유틸을 debateRoom 패키지로 이동
